+++ lib/src/lottie_image_asset.dart
@@ -1,5 +1,7 @@
 import 'dart:ui' as ui;
 
+import 'package:flutter/painting.dart';
+
 class LottieImageAsset {
   final int width;
   final int height;
@@ -7,6 +9,7 @@ class LottieImageAsset {
   final String fileName;
   final String dirName;
   ui.Image? loadedImage;
+  ImageProvider? provider;
 
   LottieImageAsset(
       {required this.width,
@@ -14,6 +17,10 @@ class LottieImageAsset {
       required this.id,
       required this.fileName,
       required this.dirName});
+
+  void dispose() {
+    provider?.evict();
+  }
 
   @override
   String toString() =>