+++ lib/src/providers/file_provider.dart
@@ -1,5 +1,7 @@
 import 'dart:ui' as ui;
+
 import 'package:path/path.dart' as p;
+
 import '../composition.dart';
 import '../lottie_image_asset.dart';
 import 'load_image.dart';
@@ -13,8 +15,10 @@ class FileLottie extends LottieProvider {
   final Object /*io.File|html.File*/ file;
 
   @override
+  String get cacheKey => 'file-${io.filePath(file)}';
+
+  @override
   Future<LottieComposition> load() async {
-    var cacheKey = 'file-${io.filePath(file)}';
     return sharedLottieCache.putIfAbsent(cacheKey, () async {
       var bytes = await io.loadFile(file);
       var composition = await LottieComposition.fromBytes(bytes,