+++ lib/src/providers/memory_provider.dart
@@ -1,7 +1,9 @@
 import 'dart:typed_data';
 import 'dart:ui' as ui;
+
 import 'package:flutter/rendering.dart';
 import 'package:path/path.dart' as p;
+
 import '../composition.dart';
 import '../lottie_image_asset.dart';
 import 'load_image.dart';
@@ -14,9 +16,11 @@ class MemoryLottie extends LottieProvider {
   final Uint8List bytes;
 
   @override
+  String get cacheKey => 'memory-${bytes.hashCode}-${bytes.lengthInBytes}';
+
+  @override
   Future<LottieComposition> load() async {
     // TODO(xha): hash the list content
-    var cacheKey = 'memory-${bytes.hashCode}-${bytes.lengthInBytes}';
     return sharedLottieCache.putIfAbsent(cacheKey, () async {
       var composition = await LottieComposition.fromBytes(bytes,
           imageProviderFactory: imageProviderFactory);